Use vuex-connect for Detail component
